+++ feincms3/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (3, 3, 5)
+VERSION = (3, 4, 0)
 __version__ = ".".join(map(str, VERSION))